Restructured scopes in better placed files.
